# Re-add the functions for division and subtraction, faster than composite functions