I would consider re-adding division and subtraction. Subtraction is 50% faster, so even though they might seem superfluous, there is some speed to be gained on these very common operations.

## Subtraction:
```
BenchmarkSub10001                 300000              5053 ns/op        15831.81 MB/s
BenchmarkSubScaleMul10001         200000              7925 ns/op        10095.07 MB/s
```

## Division:
```
BenchmarkDiv10001          50000             34982 ns/op        2287.12 MB/s
BenchmarkDivRcpMul10001    50000             40462 ns/op        1977.35 MB/s
```